Restructure load and save so that data can be imported
