+++ tests/test_huracanpy.py
@@ -2,6 +2,7 @@ import pytest
 import numpy as np
 
 import huracanpy
+from huracanpy.load._netcdf import _find_trajectory_id
 
 
 def test_load_track():
@@ -19,7 +20,7 @@ def test_load_csv():
 def test_load_netcdf():
     data = huracanpy.load(huracanpy.example_TRACK_netcdf_file)
     assert len(data.time) == 4580
-    track_id = huracanpy._tracker_specific.netcdf._find_trajectory_id(data)
+    track_id = _find_trajectory_id(data)
     assert len(track_id) == 4580
     assert len(np.unique(track_id)) == 86
